Ajuste em cooperado home, por engano criei Controller, View para AdmCooperado sem necessidade. Desculpa!
