+++ Codigo/EcoConecteWeb/EcoConecteWeb/Views/Shared/_LoginPartial.cshtml
@@ -76,7 +76,7 @@
         @if (ehCooperado)
         {
             <li class="nav-item">
-                <a class="nav-link" asp-controller="AdmCooperado" asp-action="Home" asp-route-id="@pessoaId">
+                <a class="nav-link" asp-controller="Cooperado" asp-action="Home" asp-route-id="@pessoaId">
                     <i class="bi bi-layout-text-window-reverse me-2"></i>Meu Painel
                 </a>
             </li>
